+++ daskms/columns.py
@@ -97,7 +97,9 @@ class ColumnMetadataError(Exception):
     pass
 
 
-ColumnMetadata = namedtuple("ColumnMetadata", ["shape", "dims", "chunks", "dtype"])
+ColumnMetadata = namedtuple(
+    "ColumnMetadata", ["shape", "dims", "chunks", "dtype", "keywords"]
+)
 
 
 def column_metadata(column, table_proxy, table_schema, chunks, exemplar_row=0):
@@ -122,16 +124,8 @@ def column_metadata(column, table_proxy, table_schema, chunks, exemplar_row=0):
 
     Returns
     -------
-    shape : tuple
-        Shape of column (excluding the row dimension).
-        For example :code:`(16, 4)`
-    dims : tuple
-        Dask dimension schema. For example :code:`("chan", "corr")`
-    dim_chunks : list of tuples
-        Dimension chunks. For example :code:`[chan_chunks, corr_chunks]`.
-    dtype : :class:`numpy.dtype`
-        Column data type (numpy)
-
+    column_metadata : ColumnMetadata
+        namedtuple containing Column Metadata
 
     Raises
     ------
@@ -243,7 +237,8 @@ def column_metadata(column, table_proxy, table_schema, chunks, exemplar_row=0):
             "dim_chunks '%s' do not agree." % (shape, dims, dim_chunks)
         )
 
-    return ColumnMetadata(shape, dims, dim_chunks, dtype)
+    keywords = coldesc.get("keywords", None)
+    return ColumnMetadata(shape, dims, dim_chunks, dtype, keywords)
 
 
 def dim_extents_array(dim, chunks):