+++ daskms/tests/test_dataset_schema.py
@@ -49,5 +49,6 @@ def test_unified_schema(ms):
             assert s.shape == var.shape[1:]
             assert np.dtype(s.dtype) == var.dtype
             assert isinstance(var.data, s.type)
+            assert s.attrs == var.attrs
 
     schema.to_arrow_schema()